% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\name{absolute_lineage_sizes.LTData}
\alias{absolute_lineage_sizes.LTData}
\title{Transform lineage sizes from (relative) read counts into (absolute) cells counts}
\usage{
\method{absolute_lineage_sizes}{LTData}(lt)
}
\description{
Transform lineage sizes from (relative) read counts into (absolute) cells counts
}
