% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\name{absolute_lineage_sizes}
\alias{absolute_lineage_sizes}
\title{Transform lineage sizes from (relative) read counts into (absolute) cells counts}
\usage{
absolute_lineage_sizes(lt, ...)
}
\description{
Transform lineage sizes from (relative) read counts into (absolute) cells counts
}
