% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{adapt}
\alias{adapt}
\title{Adapt a SAN model parametrization (SANP instance) to a specific dataset}
\usage{
adapt(parameterization, ...)
}
\description{
Adapt a SAN model parametrization (SANP instance) to a specific dataset
}
