% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{adapt.SANParametrization}
\alias{adapt.SANParametrization}
\title{Adapt a SAN model parametrization (SANP instance) to a specific dataset}
\usage{
\method{adapt}{SANParametrization}(parameterization, lt)
}
\description{
Adapt a SAN model parametrization (SANP instance) to a specific dataset
}
