% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{autodetect.metaparameters}
\alias{autodetect.metaparameters}
\title{Adapt a SAN model parametrization (SANP instance) to a specific dataset}
\usage{
autodetect.metaparameters(parameterization, ...)
}
\description{
Adapt a SAN model parametrization (SANP instance) to a specific dataset
}
