% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{bandwidth.matrix.default}
\alias{bandwidth.matrix.default}
\title{Find a suitable bandwidth matrix for kernel density estimation}
\usage{
\method{bandwidth.matrix}{default}(x, H = "auto")
}
\description{
Find a suitable bandwidth matrix for kernel density estimation
}
