% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{bandwidth.matrix}
\alias{bandwidth.matrix}
\title{Find a suitable bandwidth matrix for kernel density estimation}
\usage{
bandwidth.matrix(x, ...)
}
\description{
Find a suitable bandwidth matrix for kernel density estimation
}
