% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{bandwidth.matrix.SANMCMC}
\alias{bandwidth.matrix.SANMCMC}
\title{Find a suitable bandwith matrix for kernel density estimation and sampling of the posterior}
\usage{
\method{bandwidth.matrix}{SANMCMC}(sanmcmc, H = "auto")
}
\description{
Find a suitable bandwith matrix for kernel density estimation and sampling of the posterior
}
