% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\docType{data}
\name{esk2020.mcmc}
\alias{esk2020.mcmc}
\title{The MCMC results for the esk2020 dataset}
\format{
An object of class \code{SANMCMC} of length 5.
}
\usage{
esk2020.mcmc
}
\description{
The MCMC results for the esk2020 dataset
}
\keyword{datasets}
