% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\docType{data}
\name{esk2020.model}
\alias{esk2020.model}
\title{The estimated SAN model parameters for the data from Esk et al., 2020}
\format{
An object of class \code{SANModel} of length 2.
}
\usage{
esk2020.model
}
\description{
The estimated SAN model parameters for the data from Esk et al., 2020
}
\keyword{datasets}
