% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\docType{data}
\name{esk2020}
\alias{esk2020}
\title{The cerebral organoid lineage tracing dataset from Esk et al., 2020}
\format{
An object of class \code{LTData} of length 5.
}
\usage{
esk2020
}
\description{
The cerebral organoid lineage tracing dataset from Esk et al., 2020
}
\keyword{datasets}
