% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\name{estimate_reads_per_cell.LTData}
\alias{estimate_reads_per_cell.LTData}
\title{Re-estimate the number of reads per cell without relying on the total organoid sizes}
\usage{
\method{estimate_reads_per_cell}{LTData}(lt, method = "singleton_mode")
}
\description{
Re-estimate the number of reads per cell without relying on the total organoid sizes
}
