% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\name{estimate_reads_per_cell}
\alias{estimate_reads_per_cell}
\title{Re-estimate the number of reads per cell without relying on the total organoid sizes}
\usage{
estimate_reads_per_cell(lt, ...)
}
\description{
Re-estimate the number of reads per cell without relying on the total organoid sizes
}
