% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\name{estimate_sequencing_parameters.LTData}
\alias{estimate_sequencing_parameters.LTData}
\title{Estimate sequencing parameters (library size, PCR efficiency, phantom threshold)}
\usage{
\method{estimate_sequencing_parameters}{LTData}(lt, pcr_efficiency = NA, molecules = 1)
}
\description{
Estimate sequencing parameters (library size, PCR efficiency, phantom threshold)
}
