% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\name{estimate_sequencing_parameters}
\alias{estimate_sequencing_parameters}
\title{Estimate sequencing parameters (library size, PCR efficiency, phantom threshold)}
\usage{
estimate_sequencing_parameters(lt, ...)
}
\description{
Estimate sequencing parameters (library size, PCR efficiency, phantom threshold)
}
