% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_size_utilities.R
\name{fit_powerlaw_model}
\alias{fit_powerlaw_model}
\title{Fit the powerlaw model to rank-size data}
\usage{
fit_powerlaw_model(ranked, alpha = NA, r.nonzipf = c(NA_integer_, NA_integer_))
}
\arguments{
\item{ranked}{a `data.table`  with columns `sid` (sample id), `rank` (lineage size rank) and `size` (lineage size)}

\item{alpha}{the Pareto index alpha to use. By default, alpha is estimated for each sample separately.}

\item{r.nonzipf}{a two-component vector which specifies which range of ranks corresponds to "non-Zipfian" lineages}
}
\value{
a `data.table` with columns `sid` (sample id), `alpha` (Pareto index), `k` (Zipf exponent),
        `d` (Zipf intersect), `r` (smallest Zipf-goverened rank), `s` (largest Zipf-goverend lineage size).
}
\description{
Fit the powerlaw model to rank-size data
}
