% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{locate.modes.default}
\alias{locate.modes.default}
\title{Locates the posterior distribution's mode(s) using the mean-shift algorithm}
\usage{
\method{locate.modes}{default}(x, tolerance = 0.1, adjust = 1, H = "Hpi")
}
\description{
Locates the posterior distribution's mode(s) using the mean-shift algorithm
}
