% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{locate.modes}
\alias{locate.modes}
\title{Locates the distribution's mode(s) using the mean-shift algorithm}
\usage{
locate.modes(x, ...)
}
\description{
Locates the distribution's mode(s) using the mean-shift algorithm
}
