% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarystats.R
\name{locate.modes.SANMCMC}
\alias{locate.modes.SANMCMC}
\title{Locates the posterior distribution's mode(s) using the mean-shift algorithm}
\usage{
\method{locate.modes}{SANMCMC}(sanmcmc, tolerance = 0.1, adjust = 1)
}
\description{
Locates the posterior distribution's mode(s) using the mean-shift algorithm
}
