% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\docType{data}
\name{lt47.mcmc}
\alias{lt47.mcmc}
\title{The MCMC results for the LT47 dataset}
\format{
An object of class \code{SANMCMC} of length 5.
}
\usage{
lt47.mcmc
}
\description{
The MCMC results for the LT47 dataset
}
\keyword{datasets}
