% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\docType{data}
\name{lt47.model}
\alias{lt47.model}
\title{The estimated SAN model parameters for the LT47 dataset (same as esk2020.model)}
\format{
An object of class \code{SANModel} of length 2.
}
\usage{
lt47.model
}
\description{
The estimated SAN model parameters for the LT47 dataset (same as esk2020.model)
}
\keyword{datasets}
