% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lt47.ps}
\alias{lt47.ps}
\title{The estimated SAN model parameters for the LT47 dataset}
\format{
An object of class \code{list} of length 5.
}
\usage{
lt47.ps
}
\description{
The estimated SAN model parameters for the LT47 dataset
}
\keyword{datasets}
