% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\docType{data}
\name{lt47}
\alias{lt47}
\title{The LT47 cerebral organoid lineage tracing dataset}
\format{
An object of class \code{LTData} of length 5.
}
\usage{
lt47
}
\description{
The LT47 cerebral organoid lineage tracing dataset
}
\keyword{datasets}
