% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\name{LTData}
\alias{LTData}
\title{Create a lineage tracing data object}
\usage{
LTData(
  lineagesizes = NULL,
  organoidsizes = NULL,
  sequencing = NULL,
  unit = "reads",
  groups = character()
)
}
\description{
Create a lineage tracing data object
}
