% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{map.estimate}
\alias{map.estimate}
\title{Computes the MAP (maximum a-posteriori) estimate from MCMC results}
\usage{
map.estimate(x, ...)
}
\description{
Computes the MAP (maximum a-posteriori) estimate from MCMC results
}
