% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{map.estimate.SANMCMC}
\alias{map.estimate.SANMCMC}
\title{Computes the MAP (maximum a-posteriori) estimate from MCMC results}
\usage{
\method{map.estimate}{SANMCMC}(
  sanmcmc,
  ms.tolerance = 0.1,
  H.adjust = 1,
  H = "auto",
  method = "max.local.lh.avg",
  ms = NULL
)
}
\description{
Computes the MAP (maximum a-posteriori) estimate from MCMC results
}
