% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{Samples from a distribution using the Metropolis-Hastings MCMC algorithm}
\usage{
mcmc(
  llfun,
  variables,
  fixed = character(),
  blocks = NULL,
  llfun.average = 1,
  steps = NA,
  accepts = NA,
  chains,
  candidate.samples = chains,
  candidates = NULL,
  initial.samplewithreplacement = FALSE,
  initial = NULL,
  keep.history = FALSE,
  keep.initial = FALSE,
  keep.candidates = FALSE,
  acceptance.target = if ((length(variables) == 1) || directional.metropolis.gibbs) 0.44
    else 0.234,
  directional.metropolis.gibbs = FALSE,
  initial.Rproposal = 1,
  minimal.Rproposal = 0.2,
  initial.Vproposal = NULL,
  minimal.Vproposal.CV = 0.1,
  maximal.Vproposal.corr = 0.95,
  proposal.update.rate = 0.2,
  delta.ll.cutoff = 15,
  initial.ll.cutoff = NA,
  verbose = FALSE,
  extremely.verbose = FALSE
)
}
\description{
Samples from a distribution using the Metropolis-Hastings MCMC algorithm
}
