% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{model.instance}
\alias{model.instance}
\title{Convert a parameter vector to a model instance}
\usage{
model.instance(parameterization, ...)
}
\description{
Convert a parameter vector to a model instance
}
