% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{model.instance.SANParametrization}
\alias{model.instance.SANParametrization}
\title{Convert a parameter vector to a model instance}
\usage{
\method{model.instance}{SANParametrization}(parameterization, params)
}
\description{
Convert a parameter vector to a model instance
}
