% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\name{normalize_library_sizes.LTData}
\alias{normalize_library_sizes.LTData}
\title{Normalize lineage sizes within each sample group to the same sequencing library size}
\usage{
\method{normalize_library_sizes}{LTData}(lt, method = "scale")
}
\description{
Normalize lineage sizes within each sample group to the same sequencing library size
}
