% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmlapply.R
\name{parMLapply}
\alias{parMLapply}
\title{Parallel version of mapply that operates on lists}
\usage{
parMLapply(cluster, FUN, ..., .batchcount = 10 * length(cluster))
}
\description{
Parallel version of mapply that operates on lists
}
