% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\name{partial_organoid_sizes.LTData}
\alias{partial_organoid_sizes.LTData}
\title{Infer the total number of cells each sub-library of an organoid corresponds to}
\usage{
\method{partial_organoid_sizes}{LTData}(partial, whole, sublib)
}
\description{
Infer the total number of cells each sub-library of an organoid corresponds to
}
