% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\name{partial_organoid_sizes}
\alias{partial_organoid_sizes}
\title{Infer the total number of cells each sub-library of an organoid corresponds to}
\usage{
partial_organoid_sizes(partial, ...)
}
\description{
Infer the total number of cells each sub-library of an organoid corresponds to
}
