% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.SANCombinedPosterior}
\alias{plot.SANCombinedPosterior}
\title{Combine the plots from all components of a combined posterior distribution}
\usage{
\method{plot}{SANCombinedPosterior}(posterior, data, ll = ".ll_tot", plotlist = FALSE, rs_days = NULL, ...)
}
\description{
Combine the plots from all components of a combined posterior distribution
}
