% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.SANMCMC}
\alias{plot.SANMCMC}
\title{Plot posterior distribution}
\usage{
\method{plot}{SANMCMC}(
  sanmcmc,
  type = "violins",
  which = "final",
  point = NULL,
  expressions = names(sanmcmc$variables),
  extra = character(),
  ...
)
}
\description{
Plot posterior distribution
}
