% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_size_utilities.R
\name{rank_size}
\alias{rank_size}
\title{Compute rank-size table from a table with observed lineage sizes}
\usage{
rank_size(data, groups = c("day", "sid"))
}
\arguments{
\item{data}{a `data.table`}

\item{groups}{the name(s) of a set of columns of `data` which define the group within which sizes are ranked}

\item{size}{the name of a column of `data` containing the lineage sizes}
}
\value{
a `data.table` containing the grouping columns, `rank` (lineage size rank) and `size` (lineage size)
}
\description{
Compute rank-size table from a table with observed lineage sizes
}
