% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{rateslist_s0}
\alias{rateslist_s0}
\title{Convert a parameter vector to a rate list and inital cell count (s0)}
\usage{
rateslist_s0(parameterization, params)
}
\description{
Convert a parameter vector to a rate list and inital cell count (s0)
}
