% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanmodel.R
\name{ratestable.fill.na.data.frame}
\alias{ratestable.fill.na.data.frame}
\title{Fill empty (NA) cells in rates table}
\usage{
\method{ratestable.fill.na}{data.frame}(ratestable)
}
\arguments{
\item{ratestable}{a ratestable (a data.frame or data.table)}
}
\value{
a modified table without empty (NA) cells
}
\description{
Empty (NA) cells are set to the next non-empty value in the same column,
or to zero if no such value exists.
}
