% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanmodel.R
\name{ratestable.fill.na.SANModel}
\alias{ratestable.fill.na.SANModel}
\title{Fill empty (NA) cells in the model's rates table}
\usage{
\method{ratestable.fill.na}{SANModel}(model)
}
\arguments{
\item{model}{a SAN model created with `san_model()`}
}
\value{
a model whose rates table does not contain empty (NA) cells
}
\description{
Empty (NA) cells are set to the next non-empty value in the same column,
or to zero if no such value exists.
}
