% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/san_deterministic.R
\name{san_deterministic_regime}
\alias{san_deterministic_regime}
\title{Returns the name of the parameter regime of the given rates for the deterministic SAN model}
\usage{
san_deterministic_regime(rates, eps = 0.001)
}
\arguments{
\item{rates}{a names list with non-negative finite values for r_S, r_0 r_R, r_A, r_N, r_D.}
}
\value{
a character vector containing the name
}
\description{
If the goal is simply to evaluate the deterministic SAN model, `san_deterministic` should
be used which provides a user-friedly API and checks parameters for validity
}
