% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanmodel.R
\name{san_model.default}
\alias{san_model.default}
\title{Create a SANModel from a rates table and an initial number of cells}
\usage{
\method{san_model}{default}(rates, s0)
}
\description{
Create a SANModel from a rates table and an initial number of cells
}
