% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanmodel.R
\name{san_model}
\alias{san_model}
\title{Create a SANModel from a rates table and an initial number of cells}
\usage{
san_model(rates, s0)
}
\description{
Create a SANModel from a rates table and an initial number of cells
}
