% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{san_model.SANParametrization}
\alias{san_model.SANParametrization}
\title{Convert a parameter vector to a SAN model instance}
\usage{
\method{san_model}{SANParametrization}(parameterization, params, lt = NULL, s0 = NA, Tmax = NA)
}
\description{
Convert a parameter vector to a SAN model instance
}
