% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/san_model_explorer.R
\name{san_model_explorer}
\alias{san_model_explorer}
\title{Runs the SAN model explorer app}
\usage{
san_model_explorer(dataset, location.parametersets = ".")
}
\description{
Runs the SAN model explorer app
}
