% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{san_parametrization}
\alias{san_parametrization}
\title{Create SAN model parametrization}
\usage{
san_parametrization(
  ranges,
  basemodel,
  match.s0.to.data = TRUE,
  cc_days = NULL,
  rs_days = NULL,
  rs_ranks = NULL,
  sc_days = NULL
)
}
\description{
Create SAN model parametrization
}
