% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{san_posterior_combine}
\alias{san_posterior_combine}
\title{Combines multiple posteriors by summing up the likelihoods of the components}
\usage{
san_posterior_combine(...)
}
\description{
Combines multiple posteriors by summing up the likelihoods of the components
}
