% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{san_posterior_parallel}
\alias{san_posterior_parallel}
\title{Parallelizes likelihood evaluations for different parameter combinations}
\usage{
san_posterior_parallel(posterior, cluster)
}
\description{
Parallelizes likelihood evaluations for different parameter combinations
}
