% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing.R
\name{san_sequencing}
\alias{san_sequencing}
\title{Simulate sequencing}
\usage{
san_sequencing(
  lineagesizes,
  parameters,
  steps = c("aliases", "seqsim", "threshold"),
  method.seqsim = NULL
)
}
\description{
To pass the results of `san_stochastic` as the `lineagesizes` parameters, the
`t` column has to be renamed to `day`, and columns `C` containing the total number
of cells (i.e., S+A+N) and `sid` (containing an arbitrary sample id) must be added.
}
