% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/san_stochastic.R
\name{san_stochastic_steps}
\alias{san_stochastic_steps}
\title{Determine a suitable step size and steps vector for `san_stochastic_simulate_fixedrates`}
\usage{
san_stochastic_steps(rates, days, samples_per_day = 1, p_cutoff = 0.001)
}
\arguments{
\item{rates}{a names list with non-negative finite values for r_S, r_0 r_R, r_A, r_N, r_D.}

\item{days}{number of days to simulate}

\item{samples_per_day}{number of samples per day to output}

\item{p_cutoff}{the maximal allowed probability of two sequential events within one time step}
}
\value{
a list containing the step size `dt` and a vector `steps` of number of step between each sample
}
\description{
Determine a suitable step size and steps vector for `san_stochastic_simulate_fixedrates`
}
