% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanmodel.R
\name{simulate.SANModel}
\alias{simulate.SANModel}
\title{Simulate lineage tracing data}
\usage{
\method{simulate}{SANModel}(
  model,
  template,
  steps = c("aliases", "seqsim", "threshold"),
  method.seqsim = NULL,
  p_cutoff = 0.001
)
}
\description{
Simulate lineage tracing data
}
