% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{summarystats.SANMCMC}
\alias{summarystats.SANMCMC}
\title{Computes various summary statistics of the posterior distribution}
\usage{
\method{summarystats}{SANMCMC}(sanmcmc, modes, expressions = names(sanmcmc$variables))
}
\description{
Computes various summary statistics of the posterior distribution
}
