% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanmodel.R
\name{Tmax<-}
\alias{Tmax<-}
\title{Update the model's final time, i.e. the time up to which it defines rates}
\usage{
Tmax(model) <- value
}
\description{
If the final time exceeds the model's current final time, the validity interval
of the last rate set will be extended until the new final time. Otherwise, the
validity period of the rates set active at the new final time will be set to the
new final time and any later rate sets will be removed.
}
