% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanmodel.R
\name{Tmax.SANModel}
\alias{Tmax.SANModel}
\title{Return the final time, i.e. the time up to which the model defines rates}
\usage{
\method{Tmax}{SANModel}(model)
}
\description{
Return the final time, i.e. the time up to which the model defines rates
}
