% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpois.R
\name{tpois.lambda}
\alias{tpois.lambda}
\title{Compute lambda parameter for a zero-truncated Poisson distribution with mean m}
\usage{
tpois.lambda(mean)
}
\arguments{
\item{mean}{the desired mean of the zero-truncated Poisson distribution}
}
\value{
the lambda Parameter of the truncated Poisson distribution
}
\description{
Compute lambda parameter for a zero-truncated Poisson distribution with mean m
}
