% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\name{estimate_sequencing_parameters.LTData}
\alias{estimate_sequencing_parameters.LTData}
\title{Estimate sequencing parameters (library size, PCR efficiency, phantom threshold, read_per_cell)}
\usage{
\method{estimate_sequencing_parameters}{LTData}(
  lt,
  pcr_efficiency = NA,
  lineage_aliases = NA,
  molecules = 1,
  replace = FALSE
)
}
\description{
Estimate sequencing parameters (library size, PCR efficiency, phantom threshold, read_per_cell)
}
