% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdata.R
\name{normalize_library_sizes}
\alias{normalize_library_sizes}
\title{Normalize lineage sizes within each sample group to the same sequencing library size}
\usage{
normalize_library_sizes(lt, ...)
}
\description{
Normalize lineage sizes within each sample group to the same sequencing library size
}
