% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.SANPosterior}
\alias{plot.SANPosterior}
\title{Plot log-normal confidence intervals defining the posterior overlayed with model predictions}
\usage{
\method{plot}{SANPosterior}(posterior, params = NULL, llout = NULL)
}
\description{
Plot log-normal confidence intervals defining the posterior overlayed with model predictions
}
