% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{san_parametrization}
\alias{san_parametrization}
\title{Create SAN model parametrization}
\usage{
san_parametrization(
  ranges,
  basemodel,
  s0 = "autodetect",
  cc_days = "autodetect",
  rs_days = "autodetect",
  rs_ranks = "autodetect",
  sc_days = "autodetect"
)
}
\description{
Create SAN model parametrization
}
