% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{san_posterior}
\alias{san_posterior}
\title{Create a posterior distribution instance from a SAN model parametrization and a dataset}
\usage{
san_posterior(parameterization, ...)
}
\description{
Create a posterior distribution instance from a SAN model parametrization and a dataset
}
